Two table_continuous_lm() fixes: reserved covariate names + glance df.residual precision
1. resolve_covariates_argument() now rejects covariate columns
   literally named "x" or "y" with a clear spicy_invalid_input
   error. The internal model fit uses data.frame(y=outcome, x=predictor)
   and the formula y ~ x + <covariates>, so a covariate column with
   either reserved name silently shadowed the placeholder and
   produced cryptic "la réponse est apparue dans le membre de droite"
   warnings followed by an opaque "arguments inadequats" failure deep
   in the emmean math. The error now points the user to dplyr::rename()
   the offending column.

2. glance.spicy_continuous_lm_table() now keeps df.residual numeric
   instead of coercing through as.integer(). Under cluster-robust
   variance (CR2/CR3) the Satterthwaite df is genuinely fractional
   (e.g. 38.7) and may also arrive as integer-with-FP-noise (e.g.
   47.999999... very close to 48), which previously truncated to 47.
   Mirrors the broom convention for Satterthwaite-corrected models.

Adds two regression tests for (1) at the helper level (covariate
named "x", named "y") plus a glance-level test for (2) covering both
the FP-noise case and a synthetic genuinely-fractional df.

+++ R/table_continuous_lm_print.R
@@ -335,7 +335,15 @@ glance.spicy_continuous_lm_table <- function(x, ...) {
     test_type = per_outcome$test_type,
     statistic = per_outcome$statistic,
     df = as.integer(per_outcome$df1),
-    df.residual = as.integer(per_outcome$df2),
+    # `df2` is the denominator df. For classical / HC* it is an
+    # integer (`df.residual(fit)`); for CR* it is the Satterthwaite df,
+    # which is genuinely fractional (e.g. 38.7) and may also arrive as
+    # integer-but-with-FP-noise (e.g. 47.999999... very close to 48).
+    # Coercing to integer truncates both genuinely-fractional values
+    # and FP-noisy near-integers (47.999... -> 47), so keep it
+    # numeric. Mirrors the broom convention for Satterthwaite-corrected
+    # models (e.g. lmerTest::glance, afex output).
+    df.residual = as.numeric(per_outcome$df2),
     p.value = per_outcome$p.value,
     r.squared = per_outcome$r2,
     adj.r.squared = per_outcome$adj_r2,
